import/baserock_definitions: Check base URI is valid
This is actually done to make it clearer that the second parameter is
a URI definition, not a path to the output file. Output goes on stdout
in all cases. If you pass a file path as the second argument, it won't
be a valid URL so the user now gets an error.